+++ benches/buffer_mut.rs
@@ -18,6 +18,7 @@ fn buffer_mut(c: &mut Criterion) {
         use winit::platform::run_on_demand::EventLoopExtRunOnDemand;
 
         let mut evl = winit::event_loop::EventLoop::new().unwrap();
+        let context = Context::new(evl.owned_display_handle()).unwrap();
         let window = evl
             .create_window(winit::window::Window::default_attributes().with_visible(false))
             .unwrap();
@@ -28,10 +29,7 @@ fn buffer_mut(c: &mut Criterion) {
             if let winit::event::Event::AboutToWait = ev {
                 elwt.exit();
 
-                let mut surface = {
-                    let context = Context::new(elwt).unwrap();
-                    Surface::new(&context, &window).unwrap()
-                };
+                let mut surface = Surface::new(&context, &window).unwrap();
 
                 let size = window.inner_size();
                 surface