+++ examples/animation.rs
@@ -3,7 +3,7 @@ use rayon::prelude::*;
 use std::f64::consts::PI;
 use std::num::NonZeroU32;
 use web_time::Instant;
-use winit::event::{Event, KeyEvent, WindowEvent};
+use winit::event::{KeyEvent, WindowEvent};
 use winit::event_loop::{ControlFlow, EventLoop};
 use winit::keyboard::{Key, NamedKey};
 
@@ -14,31 +14,32 @@ fn main() {
     let event_loop = EventLoop::new().unwrap();
     let start = Instant::now();
 
+    let context = softbuffer::Context::new(event_loop.owned_display_handle()).unwrap();
+
     let app = winit_app::WinitAppBuilder::with_init(
         |event_loop| {
             let window = winit_app::make_window(event_loop, |w| w);
 
-            let context = softbuffer::Context::new(window.clone()).unwrap();
-
             let old_size = (0, 0);
             let frames = pre_render_frames(0, 0);
 
-            (window, context, old_size, frames)
+            (window, old_size, frames)
         },
-        |_elwft, (window, context, _old_size, _frames)| {
-            softbuffer::Surface::new(context, window.clone()).unwrap()
+        move |_elwft, (window, _old_size, _frames)| {
+            softbuffer::Surface::new(&context, window.clone()).unwrap()
         },
     )
-    .with_event_handler(move |state, surface, event, elwt| {
-        let (window, _context, old_size, frames) = state;
+    .with_event_handler(move |state, surface, window_id, event, elwt| {
+        let (window, old_size, frames) = state;
 
         elwt.set_control_flow(ControlFlow::Poll);
 
+        if window_id != window.id() {
+            return;
+        }
+
         match event {
-            Event::WindowEvent {
-                window_id,
-                event: WindowEvent::Resized(size),
-            } if window_id == window.id() => {
+            WindowEvent::Resized(size) => {
                 let Some(surface) = surface else {
                     eprintln!("Resized fired before Resumed or after Suspended");
                     return;
@@ -50,10 +51,7 @@ fn main() {
                     surface.resize(width, height).unwrap();
                 }
             }
-            Event::WindowEvent {
-                window_id,
-                event: WindowEvent::RedrawRequested,
-            } if window_id == window.id() => {
+            WindowEvent::RedrawRequested => {
                 let Some(surface) = surface else {
                     eprintln!("RedrawRequested fired before Resumed or after Suspended");
                     return;
@@ -74,26 +72,23 @@ fn main() {
                 buffer.copy_from_slice(frame);
                 buffer.present().unwrap();
             }
-            Event::AboutToWait => {
-                window.request_redraw();
-            }
-            Event::WindowEvent {
+            WindowEvent::CloseRequested
+            | WindowEvent::KeyboardInput {
                 event:
-                    WindowEvent::CloseRequested
-                    | WindowEvent::KeyboardInput {
-                        event:
-                            KeyEvent {
-                                logical_key: Key::Named(NamedKey::Escape),
-                                ..
-                            },
+                    KeyEvent {
+                        logical_key: Key::Named(NamedKey::Escape),
                         ..
                     },
-                window_id,
-            } if window_id == window.id() => {
+                ..
+            } => {
                 elwt.exit();
             }
             _ => {}
         }
+    })
+    .with_about_to_wait_handler(|state, _, _| {
+        let (window, _, _) = state;
+        window.request_redraw();
     });
 
     winit_app::run_app(event_loop, app);