+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Update to `objc2` 0.6.0.
 - Bump MSRV to Rust 1.71.
+- Make `Context` cloneable.
 - Added `Buffer::width()` and `Buffer::height()` getters.
 
 # 0.4.6